Rename TokenCacheAccessor to InMemoryTokenCacheAccessor
