+++ src/Microsoft.Identity.Client/Platforms/net45/NetDesktopPlatformProxy.cs
@@ -175,7 +175,7 @@ namespace Microsoft.Identity.Client.Platforms.net45
         /// <inheritdoc />
         public override ITokenCacheAccessor CreateTokenCacheAccessor()
         {
-            return new TokenCacheAccessor();
+            return new InMemoryTokenCacheAccessor();
         }
 
         /// <inheritdoc />
