+++ src/Microsoft.Identity.Client/Platforms/netstandard13/NetStandard13PlatformProxy.cs
@@ -29,7 +29,6 @@ using System;
 using System.Threading.Tasks;
 using Microsoft.Identity.Client.Cache;
 using Microsoft.Identity.Client.Core;
-using Microsoft.Identity.Client.Internal;
 using Microsoft.Identity.Client.PlatformsCommon.Interfaces;
 using Microsoft.Identity.Client.PlatformsCommon.Shared;
 using Microsoft.Identity.Client.UI;
@@ -141,7 +140,7 @@ namespace Microsoft.Identity.Client.Platforms.netstandard13
 
         public override ITokenCacheAccessor CreateTokenCacheAccessor()
         {
-            return new TokenCacheAccessor();
+            return new InMemoryTokenCacheAccessor();
         }
 
         protected override IWebUIFactory CreateWebUiFactory() => new WebUIFactory();
