+++ src/Microsoft.Identity.Client/PlatformsCommon/Shared/InMemoryTokenCacheAccessor.cs
@@ -23,7 +23,6 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Collections.Concurrent;
@@ -32,9 +31,16 @@ using Microsoft.Identity.Client.Cache;
 using Microsoft.Identity.Client.Cache.Items;
 using Microsoft.Identity.Client.Cache.Keys;
 
-namespace Microsoft.Identity.Client
+namespace Microsoft.Identity.Client.PlatformsCommon.Shared
 {
-    internal class TokenCacheAccessor : ITokenCacheAccessor
+    /// <summary>
+    /// Keeps the 4 token cache dictionaries in memory. Token Cache extensions 
+    /// are responsible for persistance. 
+    /// </summary>
+    /// <remarks>
+    /// Only this accessor is currently compatible with the TokenCache serialization
+    /// </remarks>
+    internal class InMemoryTokenCacheAccessor : ITokenCacheAccessor
     {
         internal readonly IDictionary<string, MsalAccessTokenCacheItem> AccessTokenCacheDictionary =
             new ConcurrentDictionary<string, MsalAccessTokenCacheItem>();
